fixed return type of all payment gateways

+++ src/DTOs/ConnectIpsValidationDTO.php
@@ -11,7 +11,7 @@ class ConnectIpsValidationDTO
 {
     private function __construct(
         private readonly string $referenceId,
-        private readonly float $transactionAmount,
+        private readonly int $transactionAmount,
     ) {}
 
     /**
